Clear messages now consistent with iOS & Booklet
Clear messages were previously implementad as JSON messages of `_type`
as `MessageClear` published on both the base MQTT topic and the `/info`
topic. However, this wasn't documented and not consistent with iOS.

To bring in line, clear messages are now published as empty payloads on
the base topic (MQTT) or simply post'ed (HTTP). Receiving an empty
message on MQTT base topic is interpreted as a clear for that device.